// MeetingManagementFront/src/app/core/models/meeting-count.ts
export class MeetingCount {
    draftMeetingsCount: number = 0;
    todayMeetingsCount: number = 0;
    upcomingMeetingsCount: number = 0;
    signatureMeetingsCount: number = 0;
    finishedMeetingsCount: number = 0;
    allMeetingsCount: number = 0;
  canceledMeetingsCount: number=0;
  attendanceMeetingsCount: number=0;
}